update a lot of stuff :)

+++ src/main/kotlin/dev/nateweisz/bytestore/node/data/NodeHeartBeat.kt
@@ -4,5 +4,5 @@ data class NodeHeartBeat(
     val nodeId: String,
     val timestamp: Long,
     val cpuUsage: Double,
-    val memoryUsage: Double
+    val memoryUsage: Long
 )
